Refactored request to separate file
