+++ src/backend/api/Fusion.Resources.Api/Controllers/Personnel/PersonnelController.cs
@@ -396,28 +396,4 @@ namespace Fusion.Resources.Api.Controllers
             }
         }
     }
-
-    public class UpdateContractPreferredMailRequest
-    {
-        public List<PersonnelPreferredContact> Personnel { get; set; } = new List<PersonnelPreferredContact>();
-
-        public class PersonnelPreferredContact
-        {
-            public Guid PersonnelId { get; set; }
-            public string? PreferredContactMail { get; set; } 
-        }
-
-        public class Validator : AbstractValidator<UpdateContractPreferredMailRequest>
-        {
-            public Validator()
-            {
-                RuleForEach(p => p.Personnel).ChildRules(c =>
-                {
-                    c.RuleFor(x => x.PreferredContactMail).NotHaveInvalidMailDomain();
-                    c.RuleFor(x => x.PreferredContactMail).IsValidEmail()
-                        .When(x => x.PreferredContactMail != null);
-                }).When(x => x.Personnel != null);
-            }
-        }
-    }
 }
